+++ requirements.txt
@@ -1,5 +1,4 @@
-telepot==12.7
-urllib3==1.26.5
+python-telegram-bot==13.7
 RPi.GPIO==0.7.0
 picamera==1.13
 gpiozero==1.5.0
